requirements.txt
streamlit==1.26.0
pyyaml
pillow
matplotlib
numpy
torchvision
pandas
pytorch